Pycodestyle

+++ 0x00-python_variable_annotations/6-sum_mixed_list.py
@@ -5,7 +5,7 @@ Has a type-annotated function sum_mixed_list.
 from typing import List, Union
 
 
-def sum_mixed_list(mxd_list: List[Union[float, int]]) -> float:
+def sum_mixed_list(mxd_list: List[Union[int, float]]) -> float:
     """
     It takes a list mxd_lst of integers and floats
     and returns their sum as a float.
@@ -14,4 +14,5 @@ def sum_mixed_list(mxd_list: List[Union[float, int]]) -> float:
     total: float = 0
     for x in mxd_list:
         total += x
+
     return total
